This one does a lot...
1) Added a bunch of documentation, mostly JavaDoc style. Should be easy to parse and export with a slight modification to my cppdoc generator.

2) Added a wall between the agent and the board. Requesting the cards in the pool will now give null values for every card that hasn't been shown yet, or an empty list if it's pre-flop. The cards which are set but not shown are now hidden behind a layer of abstraction and privacy bounds.

3) Fixed a bug with the cards Enum values. More on that below.

4) Made all references to an enumerated value, such as card IDs, suits, or stage states, referenced by their name instead of value.

5) Added in another layer of redundancy against drawing more cards from the deck than there actually are. This means if you try to draw too many, it'll stop you. This adds 1 to 2 additional cycles per draw, however I'm currently looking at branch prediction for Python to lessen this.

6) Made some changed to accomidate having more than one deck's worth of cards in a deck. I was short-sighted before, and have now corrected the issue, to my knowledge. However, this may come up again later if I missed something.

7) Fixed some inaccurate error codes.

8) Got rid of/commented out some print lines I used for debugging. If they're still in the source, they're tagged with @DEBUG , so they're easy to grep out before production.

Now, about that bug.........

Long story short, I had their IDs 1-indexed to be equal to their numerical values when modulo'd by 13. At least, that was the idea... However, I was dumb and wasn't thinking about the fact that that would mean the King's numerical value would be 0, since its ID was 13. This caused issues in a few things, but the biggest one was that the suit calculation for kings would be wrong, and the last one would crash it since there was no suit with an enum value of 4. While this would also affect things like straight calc's, I hadn't made those yet, so I got off easy.